implementação de mais alguns testes, exceptions e arquivos shell para facilitar execução.

+++ src/main/java/br/com/hellobank/service/ClienteServiceImpl.java
@@ -6,6 +6,8 @@ import javax.transaction.Transactional;
 
 import org.springframework.stereotype.Service;
 
+import br.com.hellobank.exception.ClienteNotFoundException;
+import br.com.hellobank.exception.CpfNotFoundException;
 import br.com.hellobank.model.Cliente;
 import br.com.hellobank.repository.ClienteRepository;
 
@@ -29,26 +31,26 @@ public class ClienteServiceImpl implements IClienteService {
 
 	public Cliente buscarPorId(Long id) {
 
+		if(!clienteRepository.existsById(id)) {
+			throw new ClienteNotFoundException(id);
+		}
+
 		return clienteRepository.findById(id).orElse(null);
 	}
 
 	public Cliente buscarPorCpf(String cpf) {
+		if (!clienteRepository.existsByCpf(cpf)) {
+			throw new CpfNotFoundException(cpf);
+		}
 
 		return clienteRepository.findByCpf(cpf);
-	}
-	
-	public boolean existeCpf(String cpf) {
-		
-		return clienteRepository.existsByCpf(cpf);
-	}
-	
-	public boolean existeId(Long id) {
-		
-		return clienteRepository.existsById(id);
 	}
 
 	@Transactional
 	public void deletar(Long id) {
+		if(!clienteRepository.existsById(id)) {
+			throw new ClienteNotFoundException(id);
+		}
 
 		clienteRepository.deleteById(id);
 	}
